Group fixes
Group items were never painted. Also the detection if an item was a group was 
broken so that the context menu and main menu items were never activated.

Be careful with it: creating and deleting groups works. If you have a group 
defined and start the program it will crash. Also editing a group crashes.

svn path=/trunk/KDE/kdeutils/kgpg/; revision=705122