Remove another instance of the unportable popen()
The newly introduced function can work with more than one signature at once. 
Currently this is not supported by the rest of the program but I don't see 
any benefit in introducing a new function with another limitation.

Deletion of multiple signatures at once will come when the feature freeze is 
lifted.

CCBUG:68470

svn path=/trunk/KDE/kdeutils/kgpg/; revision=709135