Fix the display of GnuPG encoded UTF8
GnuPG recodes some UTF8 characters as \xnn (where nn are hex numbers). When 
this happens to one of character of a multibyte entity (e.g. one of the bytes 
of a german umlaut) the display of this string is destroyed. We receive the 
data from K3ProcIO which does the transition from QByteArray to QString. 
QString then does UTF8 recoding, finds an invalid entity and overwrites it. 
Now you can't get the correct display even if you recode the \xnn to the 
original value.

Since I can't change K3ProcIO I copied it and modified it to do the \xnn
recoding before the translation to QString.

BUG:139276

svn path=/trunk/KDE/kdeutils/kgpg/; revision=711018