Fix list of selected items
Make sure that only visible items are in the list of selected ones. Otherwise 
the behaviour is very strange:

startup -> mark all -> refresh keys possible

expand a key -> collapse it -> mark all -> refresh keys not possible

This is because in the second example all items get selected, including the 
invisible signatures.

svn path=/trunk/KDE/kdeutils/kgpg/; revision=714045