Fix UTF8 in Options dialog
I just wanted to replace "--with-colon" with "--with-colons". But then I 
noticed that this is another place where listing the keys is duplicated and 
again the UTF8 decoding was wrong. So I moved the more general stuff into 
KgpgInterface and removed all that hand-made key listing stuff.

An additional goodie is that now the key pairs are listed first in the combo 
boxes. Since the "always encrypt with" key is usuall one of your secret ones 
this allows the user to select it much faster than before.

svn path=/trunk/KDE/kdeutils/kgpg/; revision=714651