Clean up fingerprint and key id handling
Key ids are the last 8 or 16 characters of the fingerprint. So no need to 
store this information 3 times per key with different lengths. Also don't 
store the beautified (i.e. space inserted) version of the fingerprint when 
it's only needed seldomly. Generate that version on demand.

As another benefit you should now be able to search in the complete 
fingerprint even if only the last 8 characters of them are displayed in 
keysmanager.

CCBUG:130552

svn path=/branches/work/kgpg2/; revision=751583