Completely rework signing/verifying/encrypting/decrypting of files and text
All these operations are now moved into it's own file. This brings down the 
size of kgpginterface.cpp down by one quarter. Nearly 100kB were too big for 
a single file anyway. The additional benefit is that now one file does not 
include kgpginterface.h anymore.

The code is not only moved but cleaned up. There are less member variables 
of the interface object now, the class is cleaner and the obsolete classes 
K3Process and K3ProcIO are not used anymore for this purpose. There are 
probably still some cornercases not handled correctly but that should be 
(hopefully) easier to fix now.

svn path=/branches/work/kgpg2/; revision=758465