Implement deleting a key pair by deleting the key pair
Weird, isn't it? If you delete a public key (which may or may be not a part 
of a key pair) you have to invalidate all signatures from this key. Until now 
after deleting a secret key all secret keys were simply reloaded. This was 
wrong in multiple different ways as it did not do everything expected. Also 
it is inefficient to reload multiple keys. Now that we have notifications 
between public keys and signature it's much cleaner: delete the key node if 
it is gone from keyring and that's it.

BUG:156875

This one should be really gone now. There is no chance that the key pair is 
still there if the key is deleted. At least I hope so ;)

svn path=/trunk/KDE/kdeutils/kgpg/; revision=789986