Implement filter by key trust
Until now there was an option to hide expired keys. This is the successor, 
allowing to set a level of trust and every key below this trust is hidden in 
the key manager. This includes support for hiding disabled and expired keys.
If you set it to "Active" (which is the default) you get the old behaviour 
with expired and revoked keys hidden. Set it to Full to get only keys shown 
you really trust.

BUG:104821

svn path=/trunk/KDE/kdeutils/kgpg/; revision=808244