Rework file decryption
If the decryption fails the filenames of those files are shown. The error 
message given in this case is not perfect, but because we are in string 
freeze I reused one that already existed.

Before this for every file to decrypt one KgpgLibrary object was created and 
never deleted again. Now only one object is created for every decryption job 
(which may contain several files at once) and afterwards is properly destroyed.

BUG:161931

svn path=/trunk/KDE/kdeutils/kgpg/; revision=822095