There should only be one KGpgItemModel ...
This not only saves some memory but also execution time as the list of keys will not be processed again every time a key select window is shown.

While at it the constructors of the key select windows were cleaned up a bit. Also the arguments were reordered so the default arguments can be used more often.

svn path=/trunk/KDE/kdeutils/kgpg/; revision=842387