Use --fixed-list-mode for GnuPG key output.
This is the default since GPG 2.0.10 and in general makes the life more easy. Should have been used forever.

Please test if this gives the same results as before and report back so I can backport this.

CCBUG:188473

Thanks to Niels Ole Salscheider for a patch that does the parsing of the new output format.

svn path=/trunk/KDE/kdeutils/kgpg/; revision=952350