properly handle self signature references
Those references must not search for the root node when a key was updated as those would usually happen when their parent is going to be deleted. Traversing the keychain upwards across that parent-on-deletion will cause crashes. At the end we don't need to do anything then: if the parent is going to be deleted this signature will follow shortly after so selfsigs never need to be refreshed.

This is the same as r1049727 in 4.3 branch

svn path=/trunk/KDE/kdeutils/kgpg/; revision=1049738