Do not copy the shortcut for jumping to the default key at application startup around
It looks like this is sort of a race condition when multiple instances start up at the same time. Looks like this goes bang if the first one is not initialized completely and the second one is already trying to access those values.

At the end this copying isn't really necessary at all: that shortcut can be queried when the function that needs this is really called.

CCBUG:203497

svn path=/trunk/KDE/kdeutils/kgpg/; revision=1110495