fix crash on updating keys
When a key was refreshed it emitted a signal from it's destructor so the references to it could be broken. The slot called here would in turn call a function on the object already in deletion. This reworks the notifications of reference nodes to not do such dangerous tricks in the destructor.

There was also the possibility that reference nodes that had no key attached would stop listening for new keys once the first wrong key was sent to them.

BUG:244831
FIXED-IN:4.5.0

svn path=/branches/KDE/4.5/kdeutils/kgpg/; revision=1151109