fix crash when generating a key fails before user entered password
If the GnuPG process started to generate a new key failed before the user
entered the password (e.g. because of a broken configuration file) we will use
an already deleted object.

The reason for this was opening a dialog and waiting for its result in a slot,
so the underlying object could have been deleted by the return to the event
loop. Now this is completely handled by signals and slots.

BUG:283357
FIXED-IN:4.7.3