get rid of KGpgTransaction::sendPassphrase()
This was dangerous as it could lead to access to deleted objects. It was only
called from 2 places. The one in KGpgChangePass could easily be converted to use
the new askNewPassphrase() method instead. The other one is the askPassphrase()
function which now does the internal call of this on its own. Luckily this
doesn't do too much harm as this in seldomly used as pinentry usually takes the
job of handling this.