remove the old key search
This has been disfunctional since KDE 4.1 and noone ever complained. Since
there is a working filter I don't see a great value in this anyway. Remove
this dead code and it's #warning.

If someone ever cares to submit a working patch I'll not reject this. But I
will neither invest any time in this.