clean up external actions
The files specified on the command line are stored in an internal variable of
the KGpgExternalActions objects. If another action is initiated before the last
one finished the list of files will be overwritten. This shuffles a lot of code
around to avoid blocking dialogs, encrypting will not use this variable anymore
and clean up some other parts.