track object deletion while GnuPG transactions are running
The process may quit and cause the object to get deleted while we are e.g.
waiting for a password input. Check if the object is still alive before reading
the next input.

If we detect the process to go away also close the password dialog as it can't
do anything useful then anymore.

BUG:296634
FIXED-IN:4.8.2