refactor KGpgExternalActions::slotEncryptDroppedFiles()
Until now this was a slot in a sort-of singleton object. The calls to it from
editor windows were routed through the KeysManager object to that global
object, or maybe even connected directly.

Now KGpgExternalActions::encryptFiles() is a static method, that creates an
object on the fly, which is then deleted once the target key is selected. Then
a KJob is fired that does the rest and deletes itself (as it is now).

For encrypting files the global object isn't used anymore.