fix KEY_CONSIDERED lines always resulting in TS_MSG_SEQUENCE errors
The list of fingerprints the line was checked against was now always empty,
because I gave the bad advise to move the code into KGpgTransaction without
closely looking on the other places that called it.

For the time being we now simply ignore all KEY_CONSIDERED lines if the
transaction has not explicitely set a list of fingerprints to match against.

This is the first instance of a bug (and even a regression) found by the
testcases written by Kavinda Pitiduwa Gamage.

Fixes: 68fc9165947c054423ac600b11462294d56fa06e