Fix filtering in the SortFilterProxy model used in KeysManager
This is the line-edit in the main window. This was broken while porting to
QRegularExpression, commit 25d9df5c637279b9. setFilterFixedString doesn't
set the QRegularExpression parts of QSortFilterProxyModel, this is fixed in
Qt6, for details see:
https://invent.kde.org/utilities/kgpg/-/merge_requests/14

Since this was using setFilterFixedString, use the pattern only so as to
match plain strings and not regex to keep the same behaviour.