Show window when starting app
Currently when starting the app the window isn't shown unless explicity requested.

This is poor UX since starting the app from Kickoff will result in no window being
shown, which is unexpected and confusing.

So rework the logic to always show the key manager window on launch/activation,
unless if explicitly requested to not show (with -t option, used in autostart file),
or the editor is requested (with -d option), or a file is specified.